JSON Updates from a specified URL
It's a bit crude right now, but the JsonUpdaterTask does indeed pull
the JSON from a specified URL and, if it differs from the local copy,
replaces the local copy with that from the internet.  It's still pretty
unsafe, but it works.

Need to integrate a dialog to announce a new version was found and ask
if it should replace it, and if we replace, then we need to restart to
load it (or just re-parse the JSON into nodes)